+++ pages/login.tsx
@@ -1,79 +1,15 @@
-import { gql, useMutation } from '@apollo/client';
-import { isLoggedInVar } from 'cache/common/common.cache';
 import { Button } from 'components/common/button/Button';
 import GoogleLoginButton from 'components/common/button/GoogleLoginButton';
 import { FormError } from 'components/common/FormError';
+import useLogin from 'hooks/pages/login/useLogin';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
-import { useForm } from 'react-hook-form';
-import { login, loginVariables } from 'src/__generated__/login';
 import UtilStyle from 'styles/common/Util.module.scss';
 import LoginStyle from 'styles/Login.module.scss';
 
-interface ILoginForm {
-  email: string;
-  password: string;
-}
-
-export const LOGIN_MUTATION = gql`
-  mutation login($loginInput: LoginInput!) {
-    login(input: $loginInput) {
-      ok
-      token
-      error
-    }
-  }
-`;
-
 export const Login = () => {
-  const router = useRouter();
-  if (router.query.keyword === 'true') {
-    router.push('/login');
-  }
-
-  const {
-    register,
-    getValues,
-    formState: { errors, isValid },
-    handleSubmit,
-  } = useForm<ILoginForm>({
-    mode: 'onChange',
-  });
-
-  const onSubmit = () => {
-    if (!loading) {
-      const { email, password } = getValues();
-      loginMutation({
-        variables: {
-          loginInput: { email, password },
-        },
-      });
-    }
-  };
-
-  const onCompleted = (data: login) => {
-    const {
-      login: { error, ok, token },
-    } = data;
-    if (ok && token) {
-      isLoggedInVar(true);
-      localStorage.setItem('folks-token', token);
-      if (localStorage.getItem('folks-token')) {
-        router.push('/');
-      }
-    } else {
-      isLoggedInVar(false);
-      if (error) {
-        console.error(error);
-        alert('로그인 에러가 발생했습니다.');
-      }
-    }
-  };
-
-  const [loginMutation, { data: loginMutationResult, loading, error }] =
-    useMutation<login, loginVariables>(LOGIN_MUTATION, {
-      onCompleted,
-    });
+  const { state, actions } = useLogin();
+  const { isValid, loading, loginMutationResult, errors } = state;
+  const { onSubmit, register, handleSubmit } = actions;
 
   return (
     <>
